Migrando funcoes de GestaoServices para FuncoesServicos
